Revert "Consolidate transform helpers into TransformTool"
